Feature view

+++ src/main/java/pl/kruchyxe/hellhealth/repository/CustomerRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import pl.kruchyxe.hellhealth.model.Customer;
 
 public interface CustomerRepository
-        extends JpaRepository<Customer, Integer> {
+        extends JpaRepository<Customer, Long> {
 
+    // count method is extended from JpaRepo
 }
